Reformat Export button and its function with loading all data #9
